Fix persistent errors in plugins
The new version of VVV will be running PHP 7.

The Log Deprecated Notices plugin doesn't really have an alternative, but is not actively maintained nor has it got a GH shadow with more up to date code.
This plugin has a PHP7 compatibility issue which this adjustment fixes.

Same goes for the Demo Data Creator plugin.

The fixes are made by editing the actual plugin files. This is done right after install, but also at the end of the script so as to fix already installed versions of the plugins.

This gets rids of the following errors which will show in the VVV provision logs and/or PHP error logs:
```
Deprecated: Methods with the same name as their class will not be constructors in a future version of PHP
```